added some files and docs to help future developers understand better
